+++ persistence/src/main/resources/db/migration/V1_4__set_default_action_state.sql
@@ -1,0 +1,5 @@
+ALTER TABLE "action"
+ALTER "state" SET DEFAULT 'NEW';
+
+ALTER TABLE "action"
+ALTER "state" SET NOT NULL;