+++ rest/src/main/java/de/starwit/rest/controller/ActionController.java
@@ -2,9 +2,6 @@ package de.starwit.rest.controller;
 
 import java.util.List;
 
-import jakarta.persistence.EntityNotFoundException;
-import jakarta.validation.Valid;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,10 +18,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import de.starwit.persistence.entity.ActionEntity;
-import de.starwit.service.impl.ActionService;
 import de.starwit.persistence.exception.NotificationException;
 import de.starwit.rest.exception.NotificationDto;
+import de.starwit.service.impl.ActionExecutorService;
+import de.starwit.service.impl.ActionService;
 import io.swagger.v3.oas.annotations.Operation;
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.validation.Valid;
 
 /**
  * Action RestController
@@ -39,6 +39,9 @@ public class ActionController {
     @Autowired
     private ActionService actionService;
 
+    @Autowired
+    private ActionExecutorService actionExecutorService;
+
     @Operation(summary = "Get all action")
     @GetMapping
     public List<ActionEntity> findAll() {
@@ -49,6 +52,12 @@ public class ActionController {
     @GetMapping(value = "/find-without-decision")
     public List<ActionEntity> findAllWithoutDecision() {
         return actionService.findAllWithoutDecision();
+    }
+
+    @Operation(summary = "Retry action execution")
+    @GetMapping("/retry-action-execution")
+    public void retryActionExecution() {
+        this.actionExecutorService.retryExecutionActions();
     }
 
     @Operation(summary = "Get action with id")