+++ webclient/app/src/app/assets/themes/general/ComponentTheme.js
@@ -26,7 +26,7 @@ const ComponentTheme = createTheme(ColorTheme,
             },
             MuiIconButton: {
                 defaultProps: {
-                    color: "inherit"
+                    color: "inherit",
                 }
             }
         }