+++ webclient/app/src/app/services/ActionRest.js
@@ -5,5 +5,10 @@ class ActionRest extends CrudRest {
     constructor() {
         super(window.location.pathname + "api/action");
     }
+
+    retryActionExecution() {
+        return axios.get(this.baseUrl + "/retry-action-execution");
+    }
+
 }
 export default ActionRest;