+++ webclient/app/vite.config.js
@@ -1,6 +1,5 @@
 import {defineConfig} from "vite";
 import react from "@vitejs/plugin-react";
-import {createUnplugin} from 'unplugin'
 
 export default defineConfig(({command}) => {
   if (command === "serve") {
@@ -10,12 +9,11 @@ export default defineConfig(({command}) => {
       ],
       base: "/ai-cockpit/",
       server: {
+        watch: {
+          usePolling: true
+        },
         proxy: {
-          "/ai-cockpit/api": "http://localhost:8081",
-          "/ai-cockpit/sbom-frontend.json": "http://localhost:8081",
-          "/ai-cockpit/sbom-backend.json": "http://localhost:8081",
-          "/ai-cockpit/aic-sbom-frontend.json": "http://localhost:8081",
-          "/ai-cockpit/aic-sbom-backend.json": "http://localhost:8081",
+          "/ai-cockpit/api": "http://localhost:8081"
         }
       },
     };